backend/requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.17
exceptiongroup==1.2.2
flake8==5.0.4
flake8-docstrings==1.7.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.1
pep8-naming==0.13.3
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pyTelegramBotAPI==4.14.1
pytest==7.1.2
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
pytest-subtests==0.9.0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2024.1
requests==2.32.3
snowballstemmer==2.2.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
gunicorn==20.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.3
django-filter==23.1
short-url==1.2.2
Pillow==11.1.0
tzdata==2024.2
cffi==1.17.1
cryptography==44.0.0
defusedxml==0.8.0rc2
oauthlib==3.2.2
pycparser==2.22
drf-extra-fields==3.7.0
PyJWT==2.10.1
PyYAML==6.0.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4